+++ libc/src/__support/math/cospif.h
@@ -22,7 +22,8 @@ namespace LIBC_NAMESPACE_DECL {
 
 namespace math {
 
-LIBC_INLINE static constexpr float cospif(float x) {
+LIBC_INLINE constexpr float cospif(float x) {
+  using namespace sincosf_utils_internal;
   using FPBits = typename fputil::FPBits<float>;
 
   FPBits xbits(x);