+++ libc/src/__support/math/exp10m1f16.h
@@ -30,7 +30,7 @@ namespace LIBC_NAMESPACE_DECL {
 
 namespace math {
 
-LIBC_INLINE static constexpr float16 exp10m1f16(float16 x) {
+LIBC_INLINE constexpr float16 exp10m1f16(float16 x) {
 
 #ifndef LIBC_MATH_HAS_SKIP_ACCURATE_PASS
   constexpr fputil::ExceptValues<float16, 3> EXP10M1F16_EXCEPTS_LO = {{