+++ libc/src/__support/math/tan.h
@@ -38,8 +38,7 @@ namespace tan_internal {
 using DoubleDouble = fputil::DoubleDouble;
 using Float128 = typename fputil::DyadicFloat<128>;
 
-LIBC_INLINE static double tan_eval(const DoubleDouble &u,
-                                   DoubleDouble &result) {
+LIBC_INLINE double tan_eval(const DoubleDouble &u, DoubleDouble &result) {
   // Evaluate tan(y) = tan(x - k * (pi/128))
   // We use the degree-9 Taylor approximation:
   //   tan(y) ~ P(y) = y + y^3/3 + 2*y^5/15 + 17*y^7/315 + 62*y^9/2835
@@ -77,7 +76,7 @@ LIBC_INLINE static double tan_eval(const DoubleDouble &u,
 
 #ifndef LIBC_MATH_HAS_SKIP_ACCURATE_PASS
 // Accurate evaluation of tan for small u.
-[[maybe_unused]] LIBC_INLINE static Float128 tan_eval(const Float128 &u) {
+[[maybe_unused]] LIBC_INLINE Float128 tan_eval(const Float128 &u) {
   Float128 u_sq = fputil::quick_mul(u, u);
 
   // tan(x) ~ x + x^3/3 + x^5 * 2/15 + x^7 * 17/315 + x^9 * 62/2835 +
@@ -124,7 +123,7 @@ LIBC_INLINE static double tan_eval(const DoubleDouble &u,
 
 } // namespace tan_internal
 
-LIBC_INLINE static double tan(double x) {
+LIBC_INLINE double tan(double x) {
   using namespace tan_internal;
   using namespace math::range_reduction_double_internal;
   using FPBits = typename fputil::FPBits<double>;