+++ libc/src/math/generic/tanpif16.cpp
@@ -39,7 +39,7 @@ constexpr fputil::ExceptValues<float16, N_EXCEPTS> TANPIF16_EXCEPTS{{
 #endif // !LIBC_MATH_HAS_SKIP_ACCURATE_PASS
 
 LLVM_LIBC_FUNCTION(float16, tanpif16, (float16 x)) {
-  using namespace sincosf16_internal;
+  using namespace math::sincosf16_internal;
   using FPBits = typename fputil::FPBits<float16>;
   FPBits xbits(x);